Move simple_client::Response to correct folder
