+++ lib/wcc/contentful/simple_client/response.rb
